Fix failing test
